% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmatrix-package.R
\docType{package}
\name{rsmatrix-package}
\alias{rsmatrix}
\alias{rsmatrix-package}
\title{rsmatrix: Matrices for Repeat-Sales Price Indexes}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Calculate the matrices in Shiller (1991, \doi{10.1016/S1051-1377(05)80028-2}) that serve as the foundation for many repeat-sales price indexes.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://marberts.github.io/rsmatrix}
  \item \url{https://github.com/marberts/rsmatrix}
  \item Report bugs at \url{https://github.com/marberts/rsmatrix/issues}
}

}
\author{
\strong{Maintainer}: Steve Martin \email{stevemartin041@gmail.com} (\href{https://orcid.org/0000-0003-2544-9480}{ORCID}) [copyright holder]

}
\keyword{internal}
